requirements.txt
openai>=1.0.0
chromadb>=0.4.0
pypdf>=4.0.0
python-dotenv>=1.0.0